+++ event-store-management/event-store-management-core/src/main/java/uk/gov/justice/services/eventstore/management/verification/process/CommandResultGenerator.java
@@ -5,6 +5,7 @@ import static uk.gov.justice.services.eventstore.management.CommandResult.failur
 import static uk.gov.justice.services.eventstore.management.CommandResult.success;
 
 import uk.gov.justice.services.eventstore.management.CommandResult;
+import uk.gov.justice.services.eventstore.management.commands.VerificationCommand;
 
 import java.util.List;
 import java.util.UUID;
@@ -13,6 +14,7 @@ public class CommandResultGenerator {
 
     public CommandResult createCommandResult(
             final UUID commandId,
+            final VerificationCommand verificationCommand,
             final List<VerificationResult> successfulResults,
             final List<VerificationResult> warningResults,
             final List<VerificationResult> errorResults) {
@@ -20,7 +22,8 @@ public class CommandResultGenerator {
 
         if (errorResults.isEmpty()) {
             final String message = format(
-                    "Verification of Catchup completed successfully with %d Error(s), %d Warning(s) and %d Success(es)",
+                    "%s completed successfully with %d Error(s), %d Warning(s) and %d Success(es)",
+                    verificationCommand.getName(),
                     errorResults.size(),
                     warningResults.size(),
                     successfulResults.size());
@@ -29,7 +32,8 @@ public class CommandResultGenerator {
         }
 
         final String message = format(
-                "Verification of Catchup failed with %d Error(s), %d Warning(s) and %d Success(es)",
+                "%s failed with %d Error(s), %d Warning(s) and %d Success(es)",
+                verificationCommand.getName(),
                 errorResults.size(),
                 warningResults.size(),
                 successfulResults.size());
@@ -38,6 +42,5 @@ public class CommandResultGenerator {
                 commandId,
                 message
         );
-
     }
 }